stt_service/requirements.txt
fastapi
uvicorn
sounddevice
wavio
openai-whisper
numpy
